python/pyproject.toml
[project]
name = "invoice-ai"
version = "3.1.0"
description = "Invoice Processing AI microservice (FastAPI + Nano Banana)"
requires-python = ">=3.11,<3.13"

dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.32.0",
  "pydantic>=2.9.0",
  "pydantic-settings>=2.6.0",
  "python-multipart>=0.0.12",
  "httpx>=0.27.0",
  "pillow>=11.0.0",
  "pypdfium2>=4.30.0",
  "pytesseract>=0.3.13",
  "python-dotenv>=1.0.1",
  "openai>=1.54.0",
  "anthropic>=0.39.0",
  "prometheus-client>=0.21.0",
  "structlog>=24.4.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "pytest-asyncio>=0.24.0",
  "pytest-cov>=5.0.0",
  "httpx>=0.27.0",
  "ruff>=0.7.0",
  "mypy>=1.13.0",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
exclude = ["tests*", "config*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 120
target-version = "py311"
